--- a/src/KeywordCalls.jl
+++ b/src/KeywordCalls.jl
@@ -26,8 +26,8 @@ function _kwcall(ex)
     f, args, sorted_args = esc(f), QuoteNode.(args), QuoteNode.(sort(args))
     q = quote
         KeywordCalls._call_in_default_order(::typeof($f), nt::NamedTuple{($(sorted_args...),)}) = $f(NamedTuple{($(args...),)}(nt))
-        $f(nt::NamedTuple) = KeywordCalls._call_in_default_order($f, _sort(merge((;$(defaults...)), nt)))
-        $f(; kw...) = $f(merge((;$(defaults...)), NamedTuple(kw)))
+        $f(nt::NamedTuple) = KeywordCalls._call_in_default_order($f, _sort(merge($defaults, nt)))
+        $f(; kw...) = $f(merge($defaults, NamedTuple(kw)))
     end
     return (f=f, args=args, sorted_args=sorted_args, q=q)
 end
@@ -35,8 +35,8 @@ end
 function _parse_args(args)
     # get args dropping the tail of any expressions
     _args = map(_get_arg, args)
-    # get the `key = val` defaults
-    _defaults = filter(a -> a isa Expr, args)
+    # get the `key = val` defaults as a NamedTuple
+    _defaults = @eval (;$(filter(a -> a isa Expr, args)...))
     return _args, _defaults
 end