@mileslucas 2021-05-19
>  I originally wanted the second output from _parse_args to be the default values already collected into a named tuple but I didn't know how to do that so I just threw it inside the quote returned by _kwcall.
Okay, I've addressed this in 89cec9d so the generated code is a little cleaner. I've also included tests for `@kwstruct` since it gets the benefit for free, and I've added simple documentation to the docstrings and README.