@cscherrer 2021-05-19
Great idea, thanks for the PR!

I'd like to get some better benchmarks in place, but to start there's
https://github.com/cscherrer/KeywordCalls.jl/blob/master/benchmarks.jl

Maybe we could compare times for an `f` using your approach with a `g` using standard keyword arguments? I'd guess the `merge` might add some small overhead. I'm not sure about that, but in any case it would be good to quantify.