@mileslucas 2021-05-19
Here are some benchmarks specifically targeting the `merge`-

**On master**

```julia
julia> f(nt::NamedTuple{(:a, :b, :c)}) = nt.a^3 + nt.b^2 + nt.c
f (generic function with 1 method)

julia> @kwcall f(a,b,c)
f (generic function with 3 methods)

julia> @btime f(a=1,b=2,c=3)
  0.038 ns (0 allocations: 0 bytes)
8
```

**This PR**

when defaults is an empty named tuple-
```julia
julia> f(nt::NamedTuple{(:a, :b, :c)}) = nt.a^3 + nt.b^2 + nt.c
f (generic function with 1 method)

julia> @kwcall f(a,b,c)
f (generic function with 3 methods)

julia> @btime f(a=1,b=2,c=3)
  0.037 ns (0 allocations: 0 bytes)
8
```
when merging does happen
```julia
julia> @kwcall f(a,b,c=0)
f (generic function with 3 methods)

julia> @btime f(a=1,b=2,c=3)
  0.037 ns (0 allocations: 0 bytes)
8

julia> @btime f(a=1,b=2)
  0.038 ns (0 allocations: 0 bytes)
5
```

---

so the *runtime* performance is completely unaffected by this change